Move application related types to applications.ts
