+++ packages/types/src/shared.ts
@@ -162,30 +162,6 @@ export enum TeamMembershipStates {
   Accepted,
 }
 
-/** https://discord.com/developers/docs/topics/oauth2#application-application-flags */
-export enum ApplicationFlags {
-  /** Indicates if an app uses the Auto Moderation API. */
-  ApplicationAutoModerationRuleCreateBadge = 1 << 6,
-  /** Intent required for bots in **100 or more servers** to receive 'presence_update' events */
-  GatewayPresence = 1 << 12,
-  /** Intent required for bots in under 100 servers to receive 'presence_update' events */
-  GatewayPresenceLimited = 1 << 13,
-  /** Intent required for bots in **100 or more servers** to receive member-related events like 'guild_member_add'. */
-  GatewayGuildMembers = 1 << 14,
-  /** Intent required for bots in under 100 servers to receive member-related events like 'guild_member_add'.  */
-  GatewayGuildMembersLimited = 1 << 15,
-  /** Indicates unusual growth of an app that prevents verification */
-  VerificationPendingGuildLimit = 1 << 16,
-  /** Indicates if an app is embedded within the Discord client (currently unavailable publicly) */
-  Embedded = 1 << 17,
-  /** Intent required for bots in **100 or more servers** to receive [message content](https://support-dev.discord.com/hc/en-us/articles/4404772028055) */
-  GatewayMessageContent = 1 << 18,
-  /** Intent required for bots in under 100 servers to receive [message content](https://support-dev.discord.com/hc/en-us/articles/4404772028055), found in Bot Settings */
-  GatewayMessageContentLimited = 1 << 19,
-  /** Indicates if an app has registered global application commands */
-  ApplicationCommandBadge = 1 << 23,
-}
-
 /** https://discord.com/developers/docs/resources/channel#allowed-mentions-object-allowed-mention-types */
 export enum AllowedMentionsTypes {
   /** Controls role mentions */
